apex-dynamic-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
mysql-connector-python==8.2.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
alembic==1.13.1
python-multipart==0.0.7
pyyaml==6.0.1
